UI improvements
New UIScripts "savepage" "gotopage" "rgbsabercvars" "recordcharweapons" and "applycharweapons"

SavePage: Saves the last page the user was on for a specific menu.

GoToPage: Automatically directs the player to the last page they were on for a specific menu. If there's no page, default to "Page1"

RGBSaberCvars: Workaround so characters with rgb sabers will display their saber properly in the menu.

RecordCharWeapons: Records the character's weapons (set in the menu), maximum of 4 weapons.

ApplyCharWeapons: Applies the weapons as recorded in the previous script.