[SP] Change which skin to check
Failed to realize that the lower skin may end up getting cut off. Switched to head.

[SP] Re-prioritize skin assignment

A lot of skins were getting cut off because the game was reading the pre-set skins as a three-part skin and would only register whatever was written for the legs. It'll now read the head skin, which should never be cut off.

Custom models shouldn't be affected
Custom models should be unaffected

Was initially causing problems with save games.

[SP] Re-prioritize skin assignment (more logic)

[SP] Re-prioritize skin assignment

A lot of skins were getting cut off because the game was reading the pre-set skins as a three-part skin and would only register whatever was written for the legs. It'll now read the head skin, which should never be cut off.

Was initially causing problems with save games.