Dynamic Weapons: Verify player is using weapon
We only want to switch to the next weapon model if the player is actually holding it. Otherwise the game is checking a weapon we're not even using and is returning a false result.

This was seen if the player had a blaster pistol and their next weapon wasn't a rifle. Was causing the pistol to cycle back to its original and never progress down the line.